+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Labels in the navigation can now be translated. If the navigation-specific translation key isn't set, we'll use the title translation for the repositorys index page.
-- The button to add a new model now falls back to the translation from  common.new when the button-specific translation doesn't exist before falling back to "New %{model}".
+- The button to add a new model now falls back to the translation from  `common.new` when the button-specific translation doesn't exist before falling back to `"New %{model}"`.
+- Page title for the new page can now be customized using a `uchi.repository.[name].new.title` translation. If that translation isn't present, the title falls back to `repository.[name].button.link_to_new`, and then `common.new` before ultimately falling back to `"New %{model}"`
 
 ### Fixed